Set write deadline to zero
If the HTTP server has been created with a write timeout, the hijacked
connection will have a write deadline set. Since eventsource has no
timeouts of its own, it never resets the deadline and writes to the
connection will eventually fail.

In Go 1.0.3, this is not an issue because of a bug that prevents the
correct checking of the deadline, but revision 14943, which will be
part of Go 1.1, fixes this.